% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDGrana2023}
\alias{BDGrana2023}
\title{Grana Padano}
\format{
\subsection{\code{BDGrana2023}}{

A data frame with 2,219 rows and 20 columns:
\describe{
\item{anno}{anno di registrazione del campione}
\item{numero}{identificativo del campione}
\item{matricola}{identificativo della provenienza del campione}
\item{class}{classificazione del tipo di grana}
\item{D13C}{}
\item{D15N}{}
\item{D2H}{}
\item{Li}{}
\item{Na}{}
\item{Mn}{}
\item{Fe}{}
\item{Cu}{}
\item{Se}{}
\item{Rb}{}
\item{Sr}{}
\item{Mo}{}
\item{Ba}{}
\item{Re}{}
\item{Bi}{}
\item{U}{}
}
}
}
\source{
laboratorio dell'unità traciabilità:  \url{https://cri.fmach.it/Unita-di-ricerca/Tracciabilita}
}
\usage{
BDGrana2023
}
\description{
DB completo aggiornato il 23/08/2022 per costruzione del modello di classificazione
}
\keyword{datasets}
