% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxOut}
\alias{boxOut}
\title{Outlier identification with boxplot}
\usage{
boxOut(dataset)
}
\arguments{
\item{dataset}{a dataframe of numerical variables}
}
\value{
dataframe with a variable \code{all} specifying if the observation is outlier for all variables
}
\description{
Outlier identification with boxplot
}
