% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretModelApp.R
\name{caretModelApp}
\alias{caretModelApp}
\title{Shiny WebApp for classificaion model training.}
\usage{
caretModelApp()
}
\value{
a small shiny interactive WebApp
}
\description{
Shiny WebApp for classificaion model training.
}
