% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{caretModelUi}
\alias{caretModelUi}
\title{Ui selectInput for available classification models}
\usage{
caretModelUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui selectInput for available classification models
}
