% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{cartServer}
\alias{cartServer}
\title{write the output of decsion tree trained model into the rfTrainingUi}
\usage{
cartServer(id, data)
}
\arguments{
\item{id}{module object identifier}

\item{data}{training sample to be used in the decision tree fitting process}
}
\value{
random forest fitted model written in the rfTrainingUi
}
\description{
write the output of decsion tree trained model into the rfTrainingUi
}
