% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationTraining.R
\name{classificationApp}
\alias{classificationApp}
\title{Shiny WebApp for classificaion model training.}
\usage{
classificationApp()
}
\value{
a small shiny interactive WebApp
}
\description{
Shiny WebApp for classificaion model training.
}
