% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassModelBuildingApp.R
\name{ClassModelBuildingApp}
\alias{ClassModelBuildingApp}
\title{Shiny WebApp for classificaion model building.}
\usage{
ClassModelBuildingApp()
}
\value{
a small shiny interactive WebApp
}
\description{
Shiny WebApp for classificaion model building.
}
