% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassModelPredictionApp.R
\name{ClassModelPredictionApp}
\alias{ClassModelPredictionApp}
\title{Shiny WebApp for classificaTion model prediction.}
\usage{
ClassModelPredictionApp()
}
\description{
Shiny WebApp for classificaTion model prediction.
}
