% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{classModelServer}
\alias{classModelServer}
\title{compute classification models such as Random Forest, Linear Discriminant Analysis,
K Nearest Neighbor, Support Vector Machine, Partial Least Squares and Classification
And Regression Tree}
\usage{
classModelServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{training data used by models}
}
\description{
compute classification models such as Random Forest, Linear Discriminant Analysis,
K Nearest Neighbor, Support Vector Machine, Partial Least Squares and Classification
And Regression Tree
}
