% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{classModelUi2}
\alias{classModelUi2}
\title{Ui selectInput for available classification models}
\usage{
classModelUi2(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui selectInput for available classification models
}
