% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPred.R
\name{classPrediction}
\alias{classPrediction}
\title{Predict a model given classification model and prediction sample in input}
\usage{
classPrediction(model, sample)
}
\arguments{
\item{model}{a given classification model}

\item{sample}{a sample to predict wit respect to the given classificatin model}
}
\description{
Predict a model given classification model and prediction sample in input
}
