% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPrediction.R
\name{classPredictionApp}
\alias{classPredictionApp}
\title{Shiny WebApp for classificaTion model training.}
\usage{
classPredictionApp()
}
\description{
Shiny WebApp for classificaTion model training.
}
