% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPredictionTest.R
\name{classPredTestApp}
\alias{classPredTestApp}
\title{Test or Predict a given classification model on unseen data}
\usage{
classPredTestApp()
}
\description{
Test or Predict a given classification model on unseen data
}
