% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{databaseInputServer}
\alias{databaseInputServer}
\title{Fill inputs needed into the databaseInputUi and import the postgres database table}
\usage{
databaseInputServer(id)
}
\arguments{
\item{id}{module identifier}
}
\value{
nothing
}
\description{
Fill inputs needed into the databaseInputUi and import the postgres database table
}
