% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{databaseInputUi}
\alias{databaseInputUi}
\title{Database User Interface with for inputs used to import a postgres database table}
\usage{
databaseInputUi(id)
}
\arguments{
\item{id}{module identifier}
}
\value{
a Ui with 3 selectInput object
}
\description{
Database User Interface with for inputs used to import a postgres database table
}
