% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcs.R
\name{DataCollectionApp}
\alias{DataCollectionApp}
\title{Implement our data collection system.
Retrieve data from different file sources (excel, csv, mdb, txt) and write to qlite database
Read and write from/to database.}
\usage{
DataCollectionApp()
}
\description{
Implement our data collection system.
Retrieve data from different file sources (excel, csv, mdb, txt) and write to qlite database
Read and write from/to database.
}
