% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{dataFileInput}
\alias{dataFileInput}
\title{a fileInput object selecting data for .RDS file}
\usage{
dataFileInput(id)
}
\arguments{
\item{id}{identifier for the datamodule}
}
\value{
a fileInput Ui
}
\description{
a fileInput object selecting data for .RDS file
}
