% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{dataFileServer}
\alias{dataFileServer}
\title{Server function reading/loading .RDS file for dataFileInput}
\usage{
dataFileServer(id)
}
\arguments{
\item{id}{identifier for the datamodule}
}
\value{
a fileInput Ui
}
\description{
Server function reading/loading .RDS file for dataFileInput
}
