% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgSource.R
\name{datasourceInputServer}
\alias{datasourceInputServer}
\title{Server function for dataSourceInputUi.}
\usage{
datasourceInputServer(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Server function for dataSourceInputUi.
}
