% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgSource.R
\name{datasourceInputUi}
\alias{datasourceInputUi}
\title{A Ui selectInput object selecting the type of datasource and importing data in the App.
The supported datasources are: postgresql database, csv, excel and microsof acces files.
It is the Ui for the datasourceInputServer and they are linked by id to compose a module.}
\usage{
datasourceInputUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
A Ui selectInput object selecting the type of datasource and importing data in the App.
The supported datasources are: postgresql database, csv, excel and microsof acces files.
It is the Ui for the datasourceInputServer and they are linked by id to compose a module.
}
