% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSource.R
\name{dataUi}
\alias{dataUi}
\title{A Ui selectInput object selecting the type of datasource and importing data in the App.
The supported datasources are: sqlite database and raw data file (csv, xls, xlsx, txt, tsv and mdb).
It is the Ui for the dataServer and they are linked by id to compose a module.}
\usage{
dataUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
A Ui selectInput object selecting the type of datasource and importing data in the App.
The supported datasources are: sqlite database and raw data file (csv, xls, xlsx, txt, tsv and mdb).
It is the Ui for the dataServer and they are linked by id to compose a module.
}
