% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataViz.R
\name{dataVizOutput}
\alias{dataVizOutput}
\title{Visualize overview of data}
\usage{
dataVizOutput(id, data)
}
\arguments{
\item{id}{identifier of the data module}

\item{data}{output of the dataImport}
}
\value{
some plots written in the dataUi
}
\description{
Visualize overview of data
}
