% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataViz.R
\name{dataVizUi}
\alias{dataVizUi}
\title{User interface where the data overview will be written}
\usage{
dataVizUi(id)
}
\arguments{
\item{id}{identifier for the data module}
}
\value{
a tagList object
}
\description{
User interface where the data overview will be written
}
