% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseIO.R
\name{DBioApp}
\alias{DBioApp}
\title{Implement our data collection system with 3 types of data source: postgresql database, csv/txt and Excel/Access files.
Each tabpanel uses a different data module.
Retrieve data from different file sources (excel, csv, mdb) and write to database
Read and write from/to database.}
\usage{
DBioApp()
}
\value{
a small shiny app using data module
}
\description{
Implement our data collection system with 3 types of data source: postgresql database, csv/txt and Excel/Access files.
Each tabpanel uses a different data module.
Retrieve data from different file sources (excel, csv, mdb) and write to database
Read and write from/to database.
}
