% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscanApp.R
\name{dbscanApp}
\alias{dbscanApp}
\title{single page application for dbscan outlier training}
\usage{
dbscanApp()
}
\value{
a shiny dashboard page with dbscan outlier training tools
}
\description{
single page application for dbscan outlier training
}
