% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscanInput}
\alias{dbscanInput}
\title{Build input sliderInput container for interaction with the dbscanServer}
\usage{
dbscanInput(id)
}
\arguments{
\item{id}{identifier of dbscan module}
}
\value{
sliderInput object
}
\description{
Build input sliderInput container for interaction with the dbscanServer
}
