% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscanOutput}
\alias{dbscanOutput}
\title{Write the dbscanServer outputs into the dbscanUi}
\usage{
dbscanOutput(id, dbscan, data, predicted_scores = NULL, dataclust_pred = NULL)
}
\arguments{
\item{id}{identifier of the dbscan module}

\item{dbscan}{the fitted dbscan model object}

\item{data}{data that has been used for the fitted dbsan clustering model}

\item{predicted_scores}{projected scores into the first 2 pca dimensions}

\item{dataclust_pred}{predicted cluster}
}
\value{
nothing
}
\description{
Write the dbscanServer outputs into the dbscanUi
}
