% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscanPrediction}
\alias{dbscanPrediction}
\title{Compute dbscan prediction for data in db selected by year}
\usage{
dbscanPrediction(id, test_sample, dbscanData)
}
\arguments{
\item{id}{identifier for the dbscan module}

\item{test_sample}{test sample to be predicted}

\item{dbscanData}{a list with dbscan object and dataclust datafreme resulting from dbscanServer}
}
\value{
a list with dbscan object and dataframe with original class and fitted cluster
}
\description{
Compute dbscan prediction for data in db selected by year
}
