% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscanPredictionApp.R
\name{dbscanPredictionApp}
\alias{dbscanPredictionApp}
\title{single page application for dbscan outlier test}
\usage{
dbscanPredictionApp()
}
\value{
a shiny dashboard page with dbscan outlier training tools
}
\description{
single page application for dbscan outlier test
}
