% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscanServer}
\alias{dbscanServer}
\title{Compute dbscan}
\usage{
dbscanServer(id, knn)
}
\arguments{
\item{id}{identifier for the dbscan module}

\item{knn}{object resulting from knn analysis}
}
\value{
a list with dbscan object and dataframe with original class and fitted cluster
}
\description{
Compute dbscan
}
