% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscanUi}
\alias{dbscanUi}
\title{Build Ui container for the dbscanOuput plots}
\usage{
dbscanUi(id)
}
\arguments{
\item{id}{identifier of dbscan module}
}
\value{
Ui container as tagList object for dbscanOutput plots
}
\description{
Build Ui container for the dbscanOuput plots
}
