% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbtable}
\alias{dbtable}
\title{Import a database table given a db connection, schema name and table name}
\usage{
dbtable(src_name, schema_name, table_name)
}
\arguments{
\item{src_name}{a database connection name}

\item{schema_name}{a name of schema in the database}

\item{table_name}{a name of the table to be imported}
}
\value{
a database table
}
\description{
Import a database table given a db connection, schema name and table name
}
