% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadData}
\alias{downloadData}
\title{Function server which downloads R dataframe to Excel spreadsheet when the downloadUi is clicked.
It doesn't need to take the filename through file input since it is given by default.}
\usage{
downloadData(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{the R object to be saved in a binary file when the button is clicked}
}
\description{
Function server which downloads R dataframe to Excel spreadsheet when the downloadUi is clicked.
It doesn't need to take the filename through file input since it is given by default.
}
