% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadDBtable}
\alias{downloadDBtable}
\title{Function server which downloads data from database table when the downloadUi is clicked}
\usage{
downloadDBtable(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{data to be written in an excel file whein the button is clicked}
}
\description{
Function server which downloads data from database table when the downloadUi is clicked
}
