% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadServer}
\alias{downloadServer}
\title{Function server which downloads R objects when the downloadUi is clicked}
\usage{
downloadServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{the R object to be saved in a binary file when the button is clicked}
}
\description{
Function server which downloads R objects when the downloadUi is clicked
}
