% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadUploadServer}
\alias{downloadUploadServer}
\title{Server function for downloadUploadUi}
\usage{
downloadUploadServer(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Server function for downloadUploadUi
}
