% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadUploadUi}
\alias{downloadUploadUi}
\title{A renderUi switching between downloadUi and databaseInputUi allowing the user to upload data from
file excel to db and to download a database table to file excel.
It is the Ui for the downloadUploadServer and they are linked by id to compose a module.}
\usage{
downloadUploadUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
A renderUi switching between downloadUi and databaseInputUi allowing the user to upload data from
file excel to db and to download a database table to file excel.
It is the Ui for the downloadUploadServer and they are linked by id to compose a module.
}
