% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{edaOutput2}
\alias{edaOutput2}
\title{Visualize overview of data and descriptive statistics in the edaUi}
\usage{
edaOutput2(id, data)
}
\arguments{
\item{id}{identifier of the data module}

\item{data}{output of the dataImport}
}
\value{
descriptive statistics and missing values distribution
}
\description{
Visualize overview of data and descriptive statistics in the edaUi
}
