% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{edaUi}
\alias{edaUi}
\title{User interface where for descriptive statistics}
\usage{
edaUi(id)
}
\arguments{
\item{id}{identifier for the data module}
}
\value{
a tagList object
}
\description{
User interface where for descriptive statistics
}
