% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{ellipseType}
\alias{ellipseType}
\title{a shiny selectInput object with Character specifying frame type for data ellipses.
Possible values are 'convex', 'confidence' or types supported by stat_ellipse
including one of c("t", "norm", "euclid").}
\usage{
ellipseType(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
a shiny selectInput object with Character specifying frame type for data ellipses.
Possible values are 'convex', 'confidence' or types supported by stat_ellipse
including one of c("t", "norm", "euclid").
}
