% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_to_sqlite_dsn}
\alias{file_to_sqlite_dsn}
\title{Write to SQLite Database table through Data source Name (DSN)}
\usage{
file_to_sqlite_dsn(dsn, data, tablename, overwrite = FALSE, append = TRUE)
}
\arguments{
\item{dsn}{the data source name}

\item{data}{data}

\item{tablename}{name for the created table}

\item{overwrite}{boolean value whether the table is overwritten}

\item{append}{boolean value wheter the table is updated if TRUE or created if FALSE}
}
\description{
Write to SQLite Database table through Data source Name (DSN)
}
