% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_class_na}
\alias{fit_class_na}
\title{Fit and train some classification models with argument na.action=na.omit,
which leads to rejection of cases with missing values on any required variable.}
\usage{
fit_class_na(data, catVar, model)
}
\arguments{
\item{data}{a dataset with explanatory variables and response variable}

\item{catVar}{the response variable}

\item{model}{string specifiying which classification model will be used.}
}
\value{
a trained and fitted classification model
}
\description{
Fit and train some classification models with argument na.action=na.omit,
which leads to rejection of cases with missing values on any required variable.
}
