% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_knn}
\alias{fit_knn}
\title{Fit and train a k nearest neighbors model}
\usage{
fit_knn(data, catVar, rfe, preprocess)
}
\arguments{
\item{data}{a dataset with explanatory variables and response variable}

\item{catVar}{the response variable}

\item{rfe}{logical value whether to apply variable selection with recursive feature elimination algorythm}

\item{preprocess}{logical value whether the preprocessing steps have to be done on the explanatory variables}
}
\value{
a fitted knn model
}
\description{
Fit and train a k nearest neighbors model
}
