% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_pls}
\alias{fit_pls}
\title{Train a Partial Least Squares model}
\usage{
fit_pls(data, catVar, rfe, preprocess)
}
\arguments{
\item{data}{a dataset containing explanatory variables and the categorical variable catVar}

\item{catVar}{a categorical variable to be taken as classifier}

\item{rfe}{logical value whether variable selection has to be done with recursive feature elimination algorythm}

\item{preprocess}{logical value whether the preprocessing steps have to be done on the explanatory variables}
}
\value{
a partial least squares fitted model
}
\description{
Train a Partial Least Squares model
}
