% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_rf}
\alias{fit_rf}
\title{Train a Random Forest model}
\usage{
fit_rf(data, catVar, rfe, preprocess)
}
\arguments{
\item{data}{a dataset containing explanatory variables and the categorical variable catVar}

\item{catVar}{a categorical variable to be taken as classifier}

\item{rfe}{logical value whether variable selection has to be done with recursive feature elimination algorythm}

\item{preprocess}{logical value whether the preprocessing steps have to be done on the explanatory variables}
}
\value{
an Random Forest fitted model
}
\description{
Train a Random Forest model
}
