% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphExplorationApp.R
\name{geApp}
\alias{geApp}
\title{Graphical Exploration Analysis: Explore data with some plots and download a report.}
\usage{
geApp()
}
\value{
a small shiny app using data module
}
\description{
Graphical Exploration Analysis: Explore data with some plots and download a report.
}
