% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granaPreProcessingApp.R
\name{granaPreProcessingApp}
\alias{granaPreProcessingApp}
\title{Data Pre-Processing raw data through: renaming variable names, detect and remove samples with percentage of
below Detection Limit missingness above a given percentage, fill the below Detection Limit missingness, replacement of
wrong decimal separator (the good one is supposed to be the comma ","), conversion of data type, and remove sample with
high percentage of missing values}
\usage{
granaPreProcessingApp()
}
\description{
Data Pre-Processing raw data through: renaming variable names, detect and remove samples with percentage of
below Detection Limit missingness above a given percentage, fill the below Detection Limit missingness, replacement of
wrong decimal separator (the good one is supposed to be the comma ","), conversion of data type, and remove sample with
high percentage of missing values
}
