% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_separator.R
\name{is_point_decimal_separator}
\alias{is_point_decimal_separator}
\title{check if the point exists in a numerical value wrongly formated as a string value}
\usage{
is_point_decimal_separator(name)
}
\arguments{
\item{name}{character string of numerical value}
}
\description{
check if the point exists in a numerical value wrongly formated as a string value
}
