% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnApp.R
\name{knnApp}
\alias{knnApp}
\title{shiny single page application for rfe and knn testing}
\usage{
knnApp()
}
\value{
a shiny dashboard page with some plots
}
\description{
shiny single page application for rfe and knn testing
}
