% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{knnInputServer}
\alias{knnInputServer}
\title{Fit and train a k nearest neighbors model given preprocessing inputs}
\usage{
knnInputServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{a dataset with explanatory variables and response variable}
}
\value{
a fitted knn model
}
\description{
Fit and train a k nearest neighbors model given preprocessing inputs
}
