% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{knnServer}
\alias{knnServer}
\title{select the significant variables with recursive feature selection and apply knn}
\usage{
knnServer(id, data)
}
\arguments{
\item{id}{identifier of data module}

\item{data}{a given dataframe with response variable and predictor variable}
}
\value{
a list with a dataframe and a knn fitted object
}
\description{
select the significant variables with recursive feature selection and apply knn
}
