% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{ldaServer}
\alias{ldaServer}
\title{write the output of linear discriminant analysis training model into the rfTrainingUi}
\usage{
ldaServer(id, data)
}
\arguments{
\item{id}{module object identifier}

\item{data}{training sample to be used in the random forest fitting process}
}
\value{
random forest fitted model written in the rfTrainingUi
}
\description{
write the output of linear discriminant analysis training model into the rfTrainingUi
}
