% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{load_file}
\alias{load_file}
\title{load data from files with extension xls, xlsx, csv, txt and mdb}
\usage{
load_file(name, path, sheet_name = NULL)
}
\arguments{
\item{name}{the name of the file}

\item{path}{full path of the file}

\item{sheet_name}{use only for specifying the excel sheetname or the mdb table name}
}
\description{
load data from files with extension xls, xlsx, csv, txt and mdb
}
