% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lofOut}
\alias{lofOut}
\title{Local Outlier Factor identification}
\usage{
lofOut(X, minPts)
}
\arguments{
\item{X}{A dataset}

\item{minPts}{minimum number of points to be considered given by a knn}
}
\value{
dataframe with local outlier factor scores
}
\description{
Local Outlier Factor identification
}
