% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.R
\name{mdbInput}
\alias{mdbInput}
\title{a selectInput object for MS Access database tables or Excel workbook in Input}
\usage{
mdbInput(id)
}
\arguments{
\item{id}{identifier for the datamodule}
}
\value{
a selectInput object with the list of tables stored in the given MS Access file or
list of sheetnames in a Excel workbook file
}
\description{
a selectInput object for MS Access database tables or Excel workbook in Input
}
