% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.R
\name{mdbServer}
\alias{mdbServer}
\title{Import data from a table in the MS Access database table name given by the mdbInput}
\usage{
mdbServer(id)
}
\arguments{
\item{id}{identifier for the data module}
}
\value{
a reactive dataframe given by the table name in the selectInput object created by mdbInput
}
\description{
Import data from a table in the MS Access database table name given by the mdbInput
}
