% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{orderType}
\alias{orderType}
\title{a  shiny selectInput object with character string that denotes
how to order the axes (variables) of the parallel coordinate plot}
\usage{
orderType(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
a  shiny selectInput object with character string that denotes
how to order the axes (variables) of the parallel coordinate plot
}
