% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcoord.R
\name{outplot}
\alias{outplot}
\title{plot outlier if there is some.}
\usage{
outplot(rawdata, outdata)
}
\arguments{
\item{rawdata}{the raw dataset in input.}

\item{outdata}{the dataset with logical variable}
}
\value{
a parallel coordinates
}
\description{
plot outlier if there is some.
}
