% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcoord.R
\name{parcoord_plot}
\alias{parcoord_plot}
\title{parallel coordinate plot of data with a column for cluster membership}
\usage{
parcoord_plot(dataclust, varOrder = NULL)
}
\arguments{
\item{dataclust}{dataframe with a column for cluster membership}

\item{varOrder}{method used to order variables in the horizontal axis}
}
\value{
a dbscan object
}
\description{
parallel coordinate plot of data with a column for cluster membership
}
