% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcoord.R
\name{parcoordPlotServer}
\alias{parcoordPlotServer}
\title{Server module rendering a parallel plot coordinate. It can be linked with plotUi by id.}
\usage{
parcoordPlotServer(id, data)
}
\arguments{
\item{id}{identifier of the module}

\item{data}{dataframe}
}
\description{
Server module rendering a parallel plot coordinate. It can be linked with plotUi by id.
}
