% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcoord.R
\name{parcoordUi}
\alias{parcoordUi}
\title{User Interface for parallel coordinate plot with inputs and outputs}
\usage{
parcoordUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
User Interface for parallel coordinate plot with inputs and outputs
}
