% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_biplot.R
\name{pca_biplotServer}
\alias{pca_biplotServer}
\title{Server module rendering a pca biplot. It can be linked with plotUi by id.}
\usage{
pca_biplotServer(id, pcaObject, data)
}
\arguments{
\item{id}{identifier of the module}

\item{pcaObject}{an list object with results of pca}

\item{data}{original data passed to pca. It is used to group individuals in biplot.}
}
\description{
Server module rendering a pca biplot. It can be linked with plotUi by id.
}
