% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaApp.R
\name{pcaApp}
\alias{pcaApp}
\title{shiny single page application for Principal Component Analysis}
\usage{
pcaApp()
}
\value{
a shiny dashboard page with pca outputs
}
\description{
shiny single page application for Principal Component Analysis
}
