% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViz.R
\name{plotVizApp}
\alias{plotVizApp}
\title{Graphical Exploration Analysis: Explore data with some plots and download a report.}
\usage{
plotVizApp()
}
\value{
a small shiny app using data module
}
\description{
Graphical Exploration Analysis: Explore data with some plots and download a report.
}
