% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVizServer}
\alias{plotVizServer}
\title{Server for plotVizUi in charge of fill Ui with plots in the App}
\usage{
plotVizServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{data used}
}
\description{
Server for plotVizUi in charge of fill Ui with plots in the App
}
