% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVizUi}
\alias{plotVizUi}
\title{Ui container for some plots in Graphical Exploration Analysis}
\usage{
plotVizUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Ui container for some plots in Graphical Exploration Analysis
}
