% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred.R
\name{PredApp}
\alias{PredApp}
\title{Load a test sample of data from different sources and a fitted model
then compute prediction test of the model}
\usage{
PredApp()
}
\value{
a small shiny app for random forest test and prediction
}
\description{
Load a test sample of data from different sources and a fitted model
then compute prediction test of the model
}
