% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predTest.R
\name{predTestOutput}
\alias{predTestOutput}
\title{write the desired Random Forest Test output into the rf_testUi object}
\usage{
predTestOutput(id, model, sample)
}
\arguments{
\item{id}{module object identifier}

\item{model}{a fitted model object given by knnServer, rfServer etc....}

\item{sample}{the test sample used}
}
\value{
output test results
}
\description{
write the desired Random Forest Test output into the rf_testUi object
}
