% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predTest.R
\name{predTestServer}
\alias{predTestServer}
\title{write the desired Random Forest Test output into the rf_testUi object}
\usage{
predTestServer(id, data1, data2, pred)
}
\arguments{
\item{id}{module object identifier}

\item{data1}{data read from application}

\item{data2}{data read from application}

\item{pred}{predicted data from classPredServer}
}
\description{
write the desired Random Forest Test output into the rf_testUi object
}
