% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predTest.R
\name{predTestUi}
\alias{predTestUi}
\title{Ui for random forest model test}
\usage{
predTestUi(id)
}
\arguments{
\item{id}{module object identifier}
}
\value{
a shiny tagList object with different Ui for output visualization
}
\description{
Ui for random forest model test
}
