% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{predViz}
\alias{predViz}
\title{visualizzation on the predicted cluster points onto the fitted cluster space given by output of the function clustViz}
\usage{
predViz(fviz_clust, predicted_scores, comp = c(1, 2))
}
\arguments{
\item{fviz_clust}{ggplot object given by the clustViz function}

\item{predicted_scores}{dataframe with data used to predict and the predicted cluster column}

\item{comp}{the chosen axes or dimensions}
}
\value{
a plot with a projection of predicted cluster points into a given cluster space
}
\description{
visualizzation on the predicted cluster points onto the fitted cluster space given by output of the function clustViz
}
