% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess}
\alias{preprocess}
\title{pre-processing steps used for regressors: zero-variance elimination, near zero-value elimination,
highly correlated variable elimination, centering, scaling and bagImpute imputation for missing values.}
\usage{
preprocess(data)
}
\arguments{
\item{data}{a numerical dataset containing predictor variables.}
}
\description{
pre-processing steps used for regressors: zero-variance elimination, near zero-value elimination,
highly correlated variable elimination, centering, scaling and bagImpute imputation for missing values.
}
