% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{preProcessInputServer}
\alias{preProcessInputServer}
\title{Server function for pre-processing steps given the selected options given by preProcessInputUi.
Predictors that are not numeric are ignored in the calculations}
\usage{
preProcessInputServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{dataframe}
}
\description{
Server function for pre-processing steps given the selected options given by preProcessInputUi.
Predictors that are not numeric are ignored in the calculations
}
