% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{preProcessInputUi}
\alias{preProcessInputUi}
\title{UI inputs for pre-processing steps in data analysis. The given options are:
\enumerate{
\item variable selection through recursive feature elimination
\item "nzv" excludes "near zero-variance" predictors.
\item "corr" seeks to filter out highly correlated predictors.
\item "center" subtracts the mean of the predictor's data (again from the data in x) from the predictor value
\item "scale" divides by the standard deviation.
\item "bagImpute" applies bag imputation to missing values
}}
\usage{
preProcessInputUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
UI inputs for pre-processing steps in data analysis. The given options are:
\enumerate{
\item variable selection through recursive feature elimination
\item "nzv" excludes "near zero-variance" predictors.
\item "corr" seeks to filter out highly correlated predictors.
\item "center" subtracts the mean of the predictor's data (again from the data in x) from the predictor value
\item "scale" divides by the standard deviation.
\item "bagImpute" applies bag imputation to missing values
}
}
