% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{preProcessMethods}
\alias{preProcessMethods}
\title{Server function for preProcessInputUi.}
\usage{
preProcessMethods(id)
}
\arguments{
\item{id}{module identifier}
}
\value{
a character vector with the selected methods from preProcessInputUi to be computed in the pre-processing steps.
}
\description{
Server function for preProcessInputUi.
}
