% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{RandomForestApp}
\alias{RandomForestApp}
\title{Retrieve data from different sources and train a Random Forest Model}
\usage{
RandomForestApp()
}
\value{
a small shiny WebApp
}
\description{
Retrieve data from different sources and train a Random Forest Model
}
