% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granaManipulation.R
\name{rename_data}
\alias{rename_data}
\title{Rename data variable in the grana dataset}
\usage{
rename_data(data)
}
\arguments{
\item{data}{grana dataset}
}
\description{
Rename data variable in the grana dataset
}
