% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{rfeServer}
\alias{rfeServer}
\title{Server function for rfe module. Given a boolean input through rfeUi.
it computes or skips variable selection steps applying recursive feature elimination algorithm.}
\usage{
rfeServer(id, data)
}
\arguments{
\item{id}{module identifier}

\item{data}{dataframe}
}
\description{
Server function for rfe module. Given a boolean input through rfeUi.
it computes or skips variable selection steps applying recursive feature elimination algorithm.
}
