% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{rfeUi}
\alias{rfeUi}
\title{UI selectInput for boolean value activating or not variable selection through recursive feature elimination algorithm}
\usage{
rfeUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
UI selectInput for boolean value activating or not variable selection through recursive feature elimination algorithm
}
