% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rfeVar}
\alias{rfeVar}
\title{variable selection through recursive feature elimination based on random forest}
\usage{
rfeVar(data, catVar)
}
\arguments{
\item{data}{a dataset with response variable and quantitative regressors.
If there are categorical variable they have to be transformed in dummy variables
prior to use rfe function for variable selection}

\item{catVar}{the response variable}
}
\description{
variable selection through recursive feature elimination based on random forest
}
