% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classModels.R
\name{rfServer}
\alias{rfServer}
\title{write the output of random forest training into the rfTrainingUi}
\usage{
rfServer(id, data)
}
\arguments{
\item{id}{module object identifier}

\item{data}{training sample to be used in the random forest fitting process}
}
\value{
random forest fitted model written in the rfTrainingUi
}
\description{
write the output of random forest training into the rfTrainingUi
}
