% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFtraining.R
\name{RFTrainingApp}
\alias{RFTrainingApp}
\title{Retrieve data from different sources and train a Random Forest Model}
\usage{
RFTrainingApp()
}
\value{
a small shiny WebApp
}
\description{
Retrieve data from different sources and train a Random Forest Model
}
