% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{sheet_list}
\alias{sheet_list}
\title{extract the name of sheets (workbook) or tables (mdb) in a given file .xls or .mdb}
\usage{
sheet_list(name, path)
}
\arguments{
\item{name}{the file name}

\item{path}{the full path of the file}
}
\description{
extract the name of sheets (workbook) or tables (mdb) in a given file .xls or .mdb
}
