% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLiteDCS.R
\name{SQLiteDataCollectionApp}
\alias{SQLiteDataCollectionApp}
\title{Implement our data collection system.
Retrieve data from different file sources (excel, csv, mdb, txt) and write to qlite database from DSN or file .db
Read and write from/to database.}
\usage{
SQLiteDataCollectionApp()
}
\description{
Implement our data collection system.
Retrieve data from different file sources (excel, csv, mdb, txt) and write to qlite database from DSN or file .db
Read and write from/to database.
}
