% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsReportApp.R
\name{statsReportApp}
\alias{statsReportApp}
\title{Data Exploration Analysis through descriptive statistics of variables}
\usage{
statsReportApp()
}
\description{
Data Exploration Analysis through descriptive statistics of variables
}
