% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testApp}
\alias{testApp}
\title{Load a test sample of data from different sources and a fitted model
then compute prediction test of the model}
\usage{
testApp()
}
\value{
a small shiny app for random forest test and prediction
}
\description{
Load a test sample of data from different sources and a fitted model
then compute prediction test of the model
}
