% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainingOutput}
\alias{trainingOutput}
\title{write the output of random forest fitted model into the rfTrainingUi}
\usage{
trainingOutput(id, object)
}
\arguments{
\item{id}{module identifier}

\item{object}{a fitted model object given by knnServer, rfServer etc....}
}
\description{
write the output of random forest fitted model into the rfTrainingUi
}
