% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{trainModel}
\alias{trainModel}
\title{Train and fit predictive models of classification and regression from the caret package.}
\usage{
trainModel(data, catVar, model, preprocess = NULL)
}
\arguments{
\item{data}{a dataset with explanatory variables and response variable}

\item{catVar}{the response variable}

\item{model}{A string specifying which classification or regression model to use.
Possible values are found using names(getModelInfo())}

\item{preprocess}{a character vector specifying the pre-processing methods to be used.}
}
\value{
a fitted knn model
}
\description{
Train and fit predictive models of classification and regression from the caret package.
}
