% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_to_db}
\alias{write_to_db}
\title{write to database from a file excel}
\usage{
write_to_db(
  dsn,
  filepath,
  schemaname,
  tablename,
  overwrite = FALSE,
  append = TRUE
)
}
\arguments{
\item{dsn}{the datasource name connected to database}

\item{filepath}{the file excel full path}

\item{schemaname}{the schema's name where the table is stored}

\item{tablename}{the table's name which will be updated. if not selected a new table will be created in the selected schema}

\item{overwrite}{logical value: when FALSE and append is FALSE with not selected table, new one is created}

\item{append}{logical value: if TRUE and overwrite is FALSE, the selected table will be updated}
}
\description{
write to database from a file excel
}
