% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToDB.R
\name{writeToDBServer}
\alias{writeToDBServer}
\title{Function server in order to write into postgres database when the writeToDBUi button is clicked.
Overwrite a table is not considered, we only implements updating and creating table.}
\usage{
writeToDBServer(id, dat)
}
\arguments{
\item{id}{module identifier}

\item{dat}{data to be written into the database}
}
\description{
Function server in order to write into postgres database when the writeToDBUi button is clicked.
Overwrite a table is not considered, we only implements updating and creating table.
}
