% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToDB.R
\name{writeToDBUi}
\alias{writeToDBUi}
\title{actionButton which writes into the database from an uploaded excel file}
\usage{
writeToDBUi(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
actionButton which writes into the database from an uploaded excel file
}
